Use NSMutableArray in OrdersApi fetch methods
Switched from using swift arrays to NSMutableArrays in OrdersApi
fetch methods (e.g. getOrderLineItems, etc.) since in some cases
we need to be able to modify the line items. For example, if user
is making changes to the order, we need to be able to modify
the line items array.